requirements.txt
google-generativeai
python-dotenv
pyyaml
fastapi
uvicorn[standard]
mutagen
python-multipart